refactor(onboarding): make the rollout quest opener a plain function
It called no hooks, so the `use` prefix only bought rules-of-hooks
constraints and implied state it never had.
