+++ frontend/web/components/pages/onboarding/OnboardingRolloutQuest/index.ts
@@ -1,3 +1,3 @@
 export { default } from './OnboardingRolloutQuest'
 export type { OnboardingRolloutQuestProps } from './OnboardingRolloutQuest'
-export { default as useRolloutQuest } from './useRolloutQuest'
+export { default as openRolloutQuest } from './openRolloutQuest'
